# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    }
  },
  "paths": {
    "/api/admin/orders": {
      "get": {
        "summary": "後台訂單列表",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "paid",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "user_id": {
                                "type": "string"
                              },
                              "recipient_name": {
                                "type": "string"
                              },
                              "recipient_email": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/products": {
      "get": {
        "summary": "後台商品列表",
        "tags": [
          "Admin Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "price": {
                                "type": "integer"
                              },
                              "stock": {
                                "type": "integer"
                              },
                              "image_url": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "updated_at": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登入"
          },
          "403": {
            "description": "權限不足"
          }
        }
      },
      "post": {
        "summary": "新增商品",
        "tags": [
          "Admin Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price",
                  "stock"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "image_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "商品新增成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price": {
                          "type": "integer"
                        },
                        "stock": {
                          "type": "integer"
                        },
                        "image_url": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "updated_at": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "參數錯誤"
          }
        }
      }
    },
...
    "/api/cart": {
      "get": {
        "summary": "查看購物車",
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionId": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "product_id": {
                                "type": "string"
                              },
                              "quantity": {
                                "type": "integer"
                              },
                              "product": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "price": {
                                    "type": "integer"
                                  },
                                  "stock": {
                                    "type": "integer"
                                  },
                                  "image_url": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "加入商品到購物車",
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionId": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "quantity"
                ],
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已加入購物車",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "product_id": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "參數缺失或庫存不足"
          },
          "404": {
            "description": "商品不存在"
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "summary": "從購物車建立訂單",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientName",
                  "recipientEmail",
                  "recipientAddress"
                ],
                "properties": {
                  "recipientName": {
                    "type": "string"
                  },
                  "recipientEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "recipientAddress": {
                    "type": "string"
                  },
                  "shippingMethod": {
                    "type": "string",
                    "enum": [
                      "home",
                      "convenience_store"
                    ],
                    "default": "home",
                    "description": "宅配120元，滿1500元免基本運費；超商取貨60元不適用滿額免運。"
                  },
                  "isRemote": {
                    "type": "boolean",
                    "default": false,
                    "description": "偏遠地區加收200元。"
                  },
                  "isExpress": {
                    "type": "boolean",
                    "default": false,
                    "description": "當日急件加收250元，可與偏遠費累加；附加費不免除。"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訂單建立成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "integer",
                          "description": "商品小計（元）"
                        },
                        "shipping_fee": {
                          "type": "integer",
                          "description": "含附加費的運費總額（元）"
                        },
                        "shipping_method": {
                          "type": "string",
                          "description": "home 或 convenience_store"
                        },
                        "is_remote": {
                          "type": "integer",
                          "description": "偏遠地區旗標（0或1）"
                        },
                        "is_express": {
                          "type": "integer",
                          "description": "急件旗標（0或1）"
                        },
                        "total_amount": {
                          "type": "integer",
                          "description": "商品小計加運費（元）"
                        },
                        "status": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "購物車為空、庫存不足、收件資訊或配送條件無效"
          }
        }
      },
      "get": {
        "summary": "自己的訂單列表",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "subtotal": {
                                "type": "integer",
                                "description": "商品小計（元）"
                              },
                              "shipping_fee": {
                                "type": "integer",
                                "description": "含附加費的運費總額（元）"
                              },
                              "shipping_method": {
                                "type": "string",
                                "description": "home 或 convenience_store"
                              },
                              "is_remote": {
                                "type": "integer",
                                "description": "偏遠地區旗標（0或1）"
                              },
                              "is_express": {
                                "type": "integer",
                                "description": "急件旗標（0或1）"
                              },
                              "total_amount": {
                                "type": "integer",
                                "description": "商品小計加運費（元）"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "summary": "取得商品列表",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "products": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "price": {
                                "type": "integer"
                              },
                              "stock": {
                                "type": "integer"
                              },
                              "image_url": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              },
                              "updated_at": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer"
                            },
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}